combine zebra setup plans
