+++ src/artemis/device_setup_plans/setup_zebra.py
@@ -1,6 +1,8 @@
 import bluesky.plan_stubs as bps
 from dodal.devices.zebra import (
     DISCONNECT,
+    IN3_TTL,
+    IN4_TTL,
     OR1,
     PC_PULSE,
     TTL_DETECTOR,
@@ -83,5 +85,23 @@ def setup_zebra_for_rotation(
     yield from bps.abs_set(zebra.output.out_pvs[TTL_XSPRESS3], DISCONNECT, group=group)
     yield from bps.abs_set(zebra.output.pulse_1_input, DISCONNECT, group=group)
 
+    if wait:
+        bps.wait(group)
+
+
+def setup_zebra_for_fgs(zebra: Zebra, group="setup_zebra_for_fgs", wait=False):
+    yield from bps.abs_set(zebra.output.out_pvs[TTL_DETECTOR], IN3_TTL, group=group)
+    yield from bps.abs_set(zebra.output.out_pvs[TTL_SHUTTER], IN4_TTL, group=group)
+    yield from bps.abs_set(zebra.output.out_pvs[TTL_XSPRESS3], DISCONNECT, group=group)
+    yield from bps.abs_set(zebra.output.pulse_1_input, DISCONNECT, group=group)
+    if wait:
+        bps.wait(group)
+
+
+def set_zebra_shutter_to_manual(
+    zebra: Zebra, group="set_zebra_shutter_to_manual", wait=False
+):
+    yield from bps.abs_set(zebra.output.out_pvs[TTL_DETECTOR], PC_PULSE, group=group)
+    yield from bps.abs_set(zebra.output.out_pvs[TTL_SHUTTER], OR1, group=group)
     if wait:
         bps.wait(group)
