+++ src/artemis/experiment_plans/fast_grid_scan_plan.py
@@ -24,7 +24,7 @@ from dodal.i03 import (
 )
 
 import artemis.log
-from artemis.device_setup_plans.setup_zebra_for_fgs import (
+from artemis.device_setup_plans.setup_zebra import (
     set_zebra_shutter_to_manual,
     setup_zebra_for_fgs,
 )
